apps/api/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Data validation
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Database
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Task queue
celery[redis]==5.4.0
redis==5.2.0

# HTTP & parsing
httpx==0.28.1
lxml==5.3.0
python-multipart==0.0.12

# AWS / storage
boto3==1.35.68

# Auth
python-jose[cryptography]==3.3.0
svix==1.39.0

# Artwork validation
Pillow==10.4.0

# MusicBrainz enrichment
musicbrainzngs==0.7.1

# PDF report generation
reportlab>=4.0.0

# Stripe billing
stripe>=10.0.0


# Audio analysis
ffmpeg-python==0.2.0
soundfile==0.12.1
pyloudnorm==0.1.1
pyacoustid==1.3.0
numpy>=1.26.0

# Fraud screening (optional faster Levenshtein at scale)
rapidfuzz==3.10.1

# Seed / tooling
pyyaml==6.0.2

# DDEX Workbench SDK (installed from packages/ddex-engine/packages/python-sdk)
# pip install -e packages/ddex-engine/packages/python-sdk
# When DDEX_WORKBENCH_API_KEY is set, remote schema/Schematron validation is enabled.
# Leave unset for local-only validation (well-formedness + structural checks).

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
aiosqlite>=0.20.0  # for async SQLite in tests
